fix renderer and pixel data architecture (u8 to u32) + bytemuck addition

+++ copper/src/renderer/test_components_renderer.rs
@@ -49,7 +49,7 @@ pub struct Transform{
 pub struct Texture{
     pub width: u32,
     pub height: u32,
-    pub pixel_data: Vec<u8>, //rgba 0-255 
+    pub pixel_data: Vec<u32>, //rgba 0-255 
 }
 
 impl Component for Texture {
